packbeam_api (atomvm_packbeam v0.7.2)

A library used to generate an AtomVM AVM file from a set of files (beam files, previously built AVM files, or even arbitrary data files).

Types

avm_element/0
(opaque)

-opaque avm_element()

avm_element_name/0

-type avm_element_name() :: string().

options/0

-type options() ::
    #{prune => boolean(),
      start_module => module() | undefined,
      application_module => module() | undefined,
      include_lines => boolean()}.

path/0

-type path() :: string().

Functions

create(OutputPath, InputPaths)

-spec create(OutputPath :: path(), InputPaths :: [path()]) -> ok | {error, Reason :: term()}.

Create an AVM file.

Equivalent to create(OutputPath, InputPaths, DefaultOptions)
where DefaultOptions is #{ prune => false, start_module => undefined, application_module => undefined, include_lines => false }

create(OutputPath, InputPaths, Options)

-spec create(OutputPath :: path(), InputPaths :: [path()], Options :: options()) ->
          ok | {error, Reason :: term()}.

Create an AVM file.
This function will create an AVM file at the location specified in OutputPath, using the input files specified in InputPaths.

create(OutputPath, InputPaths, Prune, StartModule)

This function is deprecated. This function is deprecated. Use create/3 instead..

-spec create(OutputPath :: path(),
       InputPaths :: [path()],
       Prune :: boolean(),
       StartModule :: module() | undefined) ->
          ok | {error, Reason :: term()}.

create(OutputPath, InputPaths, ApplicationModule, Prune, StartModule)

This function is deprecated. This function is deprecated. Use create/3 instead..

-spec create(OutputPath :: path(),
       InputPaths :: [path()],
       ApplicationModule :: module() | undefined,
       Prune :: boolean(),
       StartModule :: module() | undefined) ->
          ok | {error, Reason :: term()}.

delete(OutputPath, InputPath, AVMElementNames)

-spec delete(OutputPath :: path(), InputPath :: path(), AVMElementNames :: [avm_element_name()]) ->
          ok | {error, Reason :: term()}.

Delete selected elements of an AVM file.
This function will delete elements of an AVM file at the location specified in InputPath, specified by the supplied list of names. The output AVM file is written to OutputPath, which may be the same as InputPath.

extract(InputPath, AVMElementNames, OutputDir)

-spec extract(InputPath :: path(), AVMElementNames :: [avm_element_name()], OutputDir :: path()) ->
           ok | {error, Reason :: term()}.

Extract all or selected elements from an AVM file.
This function will extract elements of an AVM file at the location specified in InputPath, specified by the supplied list of names. The elements from the input AVM file will be written into the specified output directory, creating any subdirectories if the AVM file elements contain path information.

get_element_data(AVMElement)

-spec get_element_data(AVMElement :: avm_element()) -> binary().

Return AVM element data.

get_element_module(AVMElement)

-spec get_element_module(AVMElement :: avm_element()) -> module() | undefined.

Return AVM element module, if the element is a BEAM file.

get_element_name(AVMElement)

-spec get_element_name(AVMElement :: avm_element()) -> avm_element_name().

Return the name of the element.

is_beam(AVMElement)

-spec is_beam(AVMElement :: avm_element()) -> boolean().

Indicates whether the AVM file element is a BEAM file.

is_entrypoint(AVMElement)

-spec is_entrypoint(AVMElement :: avm_element()) -> boolean().

Indicates whether the AVM file element is an entrypoint.

list(InputPath)

-spec list(InputPath :: path()) -> [avm_element()].

List the contents of an AVM file.
This function will list the contents of an AVM file at the location specified in InputPath.